+++ mail/mail-common/src/tests/actions/available_actions/label_as_action.rs
@@ -17,6 +17,7 @@ macro_rules! lid {
         label_id: 0.into(),
         name: Default::default(),
         color: Default::default(),
+        order: Default::default(),
         is_selected: Some(false)
     }
 ]; "TEST2: single label, not selected")]
@@ -25,6 +26,7 @@ macro_rules! lid {
         label_id: 0.into(),
         name: Default::default(),
         color: Default::default(),
+        order: Default::default(),
         is_selected: Some(true)
     }
 ]; "TEST3: single label, selected")]
@@ -35,6 +37,7 @@ macro_rules! lid {
         label_id: 0.into(),
         name: Default::default(),
         color: Default::default(),
+        order: Default::default(),
         is_selected: Some(true),
 }]; "TEST4: all selected")]
 #[test_case(
@@ -44,6 +47,7 @@ macro_rules! lid {
             label_id: 0.into(),
             name: Default::default(),
             color: Default::default(),
+            order: Default::default(),
             is_selected: Some(false),
 }]; "TEST5: none selected")]
 #[test_case(
@@ -55,8 +59,39 @@ macro_rules! lid {
         label_id: 0.into(),
         name: "name".into(),
         color: Default::default(),
+        order: Default::default(),
         is_selected: None,
 }]; "TEST6: partially selected")]
+#[test_case(
+    &[
+        label!(local_id: lid!(1), name: "C".to_string(), display_order: 3),
+        label!(local_id: lid!(2), name: "A".to_string(), display_order: 1),
+        label!(local_id: lid!(3), name: "B".to_string(), display_order: 2),
+    ],
+    |_| false,
+    &[
+        LabelAsAction {
+            label_id: 2.into(),
+            name: "A".into(),
+            color: Default::default(),
+            order: 1,
+            is_selected: Some(false),
+        },
+        LabelAsAction {
+            label_id: 3.into(),
+            name: "B".into(),
+            color: Default::default(),
+            order: 2,
+            is_selected: Some(false),
+        },
+        LabelAsAction {
+            label_id: 1.into(),
+            name: "C".into(),
+            color: Default::default(),
+            order: 3,
+            is_selected: Some(false),
+        },
+    ]; "TEST7: labels are sorted by order")]
 pub fn test_is_selected(
     labels: &[Label],
     is_selected: impl Fn(&Label) -> bool,